pyproject.toml
[project]
name = "clauserec"
version = "0.1.0"
description = "CLAUSEREC: A Contract Clause Recommendation Framework"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "faiss-cpu>=1.7.4",
    "sqlalchemy>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "pytest>=7.4.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.4.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.hatch.build.targets.wheel]
packages = ["src"]
